brain/requirements.txt
fastapi
uvicorn[standard]
langchain
pydantic
websockets
python-dotenv
openai
pyperclip
